+++ resources/lang/fr/translations.php
@@ -5,7 +5,7 @@ return [
     'help' => 'Aide',
     'open-documentation' => 'Ouvrir la documentation',
     'close' => 'Fermer',
-    'back-to-default-panel' => 'Retour',
+    'back' => 'Retour',
     'url-copied' => 'URL copiée dans votre presse-papier.',
     'code-copied' => 'Code copié dans votre presse-papier.',
 ];