+++ resources/lang/nl/translations.php
@@ -5,7 +5,7 @@ return [
     'help' => 'Help',
     'open-documentation' => 'Open documentatie',
     'close' => 'Sluit',
-    'back-to-default-panel' => 'Terug',
+    'back' => 'Terug',
     'url-copied' => 'De URL is gekopieerd naar je klembord.',
     'code-copied' => 'De code is gekopieerd naar je klembord.',
 ];