+++ resources/views/livewire/help-menu.blade.php
@@ -1,5 +1,6 @@
 @php
     use Filament\Facades\Filament;
+    use Filament\View\PanelsRenderHook;
     use Guava\FilamentKnowledgeBase\Facades\KnowledgeBase;
 
     $plugin = KnowledgeBase::plugin();
@@ -8,11 +9,15 @@
     $hasModalPreviews = $companion->hasModalPreviews();
     $hasSlideOverPreviews = $companion->hasSlideOverPreviews();
     $hasModalTitleBreadcrumbs = $companion->hasModalTitleBreadcrumbs();
+    $hasOpenDocumentationButton = ! $companion->shouldDisableOpenDocumentationButton();
     $target = $companion->shouldOpenKnowledgeBasePanelInNewTab() ? '_blank' : '_self';
     $articleClass = $plugin->getArticleClass();
+
+    $needsTopbarSpacing = $companion->getHelpMenuRenderHook() === PanelsRenderHook::TOPBAR_END;
 @endphp
 
 <div @class([
+    'ms-4' => $needsTopbarSpacing,
     'hidden' => empty($documentation),
 ])
 >
@@ -55,11 +60,13 @@
                             {!! $documentable->getSimpleHtml() !!}
                         </x-filament-knowledge-base::content>
                         <x-slot name="footerActions">
-                            <x-filament::button tag="a"
-                                                :href="$documentable->getUrl()"
-                                                :target="$target">
-                                {{ __('filament-knowledge-base::translations.open-documentation') }}
-                            </x-filament::button>
+                            @if($hasOpenDocumentationButton)
+                                <x-filament::button tag="a"
+                                                    :href="$documentable->getUrl()"
+                                                    :target="$target">
+                                    {{ __('filament-knowledge-base::translations.open-documentation') }}
+                                </x-filament::button>
+                            @endif
                             <x-filament::button color="gray"
                                                 x-on:click.prevent="$dispatch('close-modal', { id: '{{$documentable->getId()}}' })">