+++ CHANGELOG.md
@@ -1,3 +1,54 @@
 # Changelog
 
+## [3.3.0](https://github.com/GuavaCZ/filament-knowledge-base/compare/3.2.0...3.3.0) (2026-07-26)
+
+
+### Features
+
+* allow hiding the open documentation button in modal previews ([3a9acd0](https://github.com/GuavaCZ/filament-knowledge-base/commit/3a9acd02e59833344df1184c1196266fac0a5b61)), closes [#78](https://github.com/GuavaCZ/filament-knowledge-base/issues/78)
+
+
+### Bug Fixes
+
+* accept a numeric string cache ttl ([8f6c118](https://github.com/GuavaCZ/filament-knowledge-base/commit/8f6c1180903a5b80747564e4000dd385e68ded1b))
+* add spacing between the help menu and the user menu ([19bba43](https://github.com/GuavaCZ/filament-knowledge-base/commit/19bba43ed9e6fe2e56b58acfccf4e664a454f886)), closes [#92](https://github.com/GuavaCZ/filament-knowledge-base/issues/92)
+* cache rendered markdown as plain data, not the content object ([2c9c359](https://github.com/GuavaCZ/filament-knowledge-base/commit/2c9c359021e45985c3e0ebe247efcff0928a35f4)), closes [#110](https://github.com/GuavaCZ/filament-knowledge-base/issues/110)
+* normalize the docs path across platforms ([f50526f](https://github.com/GuavaCZ/filament-knowledge-base/commit/f50526f1ec76aa91c5224c18030c47771d29a4a2)), closes [#104](https://github.com/GuavaCZ/filament-knowledge-base/issues/104)
+* use the Filament v5 sidebar item class in the anchors component ([b74e3cf](https://github.com/GuavaCZ/filament-knowledge-base/commit/b74e3cf24fe1c3e7cd50e76b9ab88f9cbe85dda4))
+
+
+### Style
+
+* remove unused import ([8add914](https://github.com/GuavaCZ/filament-knowledge-base/commit/8add914b621292b17c61f83cc6f79062cbefa657))
+
+
+### Tests
+
+* assert the packaged blade views compile ([ddf0772](https://github.com/GuavaCZ/filament-knowledge-base/commit/ddf0772d4555f703f6b82c5faf744765a989edee))
+
+
+### Chore
+
+* drop debug logging from the modals component ([7958580](https://github.com/GuavaCZ/filament-knowledge-base/commit/7958580aa638d3aef66252fa9f0ff8b24d13f35a))
+
+## [3.2.0](https://github.com/GuavaCZ/filament-knowledge-base/compare/3.1.0...3.2.0) (2026-07-26)
+
+
+### Features
+
+* add configurable back button to knowledge base panel ([efdcc9d](https://github.com/GuavaCZ/filament-knowledge-base/commit/efdcc9d083a0c5bddf7636c39495a270cdc9dc87))
+
+
+### Bug Fixes
+
+* panel root redirect, type safety and dead code cleanup ([8c21e52](https://github.com/GuavaCZ/filament-knowledge-base/commit/8c21e52df221dab3df1b427a16fa6e696c65aed5))
+
+
+### Chore
+
+* add testbench workbench for local development ([256ee0d](https://github.com/GuavaCZ/filament-knowledge-base/commit/256ee0dd80f44fcc5054c515b05a5978a26b9876))
+* switch from semantic-release to release-please ([2b364b4](https://github.com/GuavaCZ/filament-knowledge-base/commit/2b364b4c7dcf2690ea188c341bb2ba922a3aa345))
+
+## Changelog
+
 All notable changes to `filament-knowledge-base` will be documented in this file.